ICE Memory mode changes
SEARCHING_H state has been added to allow enemies to search for the
player to attempt to attack them. More testing is required, but one
issue that resulted is that all ice in a room turn red regardless of
whether it notices the player

This might require changing "Mark all as hostile" to not be a default
reaction to one ICE being hostile.