Tenatively completed ICE Memory modified
The issues with ICE Memory mode have largely been resolved, pending
further testing and refinement. ICE will stay hostile, even if the
player leaves the node when it is smoked. ICE that are lost this way
will enter a hostile searching state, and will skip the query phase if
the player is discovered. Enemies will immediately raise an alarm if
they get a chance, making trying to ditch ICE this way risky.