Tweaks to program tables, other items
Adjusted the program/source code drop table to be dependent on the
system level- the new formula is Roll(100) + 45 - (system level * 3).
This will ensure that early levels will be more likely to have useful
programs, and higher levels will have programs at the same level or
lower (and effectively caps the program level at 20).

Changed code that made response ICE start searching even if the alert
level is yellow or green

Added function to the external alarm to allow backdoor creation without
risk of triggering an alert. May change some of the flavor text about
this later, but this allows Backdoor Creation (N) contracts to be
fulfilled without a flat risk of failure with no means to mitigate it.

Added some brackets in some if statements near ICE generation code.